requirements.txt
python-telegram-bot>=20.0
yt-dlp>=2024.0.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
